+++ docs/release-notes.rst
@@ -3,13 +3,26 @@ Release Notes
 ==================================
 
 *************
+Version 0.6.5
+*************
+Release Date xx/xx/xx
+
+New Features
+############
+
+Improvements
+############
+- DMM current mode functions now raise an exception when an incompatible port and range combination is used. The range and port parameters are now required parameters.
+
+
+*************
 Version 0.6.4
 *************
 Release Date 14/01/25
 
 New Features
 ############
-- DMM drivers now have a new function to set NPLC (Number of Power Line Cycles) for the DMM.
+- DMM drivers now have a new function to set NPLC (Number of Powr Line Cycles) for the DMM.
 - DMM drivers now have a new function to use the DMM's internal statistics function to take multiple measurements and return the mean, minimum and maximum values.
 
 Improvements